Port PageTemplateDialog
