+++ src/core/gui/menus/popoverMenus/PageTypeSelectionPopover.cpp
@@ -20,9 +20,10 @@
 #include "util/i18n.h"         // for _
 #include "util/raii/GVariantSPtr.h"
 
+#include "PageTypeSelectionPopoverGridHelper.h"
+
 
 namespace {
-static constexpr auto PAGE_TYPES_PER_ROW = 4;
 static constexpr auto G_ACTION_NAMESPACE = "win.";
 static constexpr auto PAGETYPE_SELECTION_ACTION_NAME = "select-page-type-of-new-page";
 static constexpr auto ORIENTATION_SELECTION_ACTION_NAME = "select-page-orientation-of-new-page";
@@ -45,18 +46,6 @@ GtkWidget* createEntryWithoutPreview(const char* label, size_t entryNb, const st
     return button;
 }
 
-GtkGrid* createEmptyGrid() {
-    GtkGrid* grid = GTK_GRID(gtk_grid_new());
-
-    gtk_grid_set_column_homogeneous(grid, true);
-    gtk_grid_set_row_homogeneous(grid, true);
-    gtk_grid_set_column_spacing(grid, 10);
-    gtk_grid_set_row_spacing(grid, 10);
-    gtk_widget_set_margin_start(GTK_WIDGET(grid), 10);
-    gtk_widget_set_margin_end(GTK_WIDGET(grid), 10);
-
-    return grid;
-}
 auto getInitiallySelectedPaperSize(const Settings* settings) -> std::optional<PaperSize> {
     if (settings) {
         return settings->getPageTemplateSettings().isCopyLastPageSize() ?
@@ -76,56 +65,6 @@ auto createPageSizeChangedGAction() -> GSimpleAction* {
     return g_simple_action_new(PAGESIZE_CHANGED_ACTION_NAME, nullptr);
 }
 
-/**
- * @brief Create a togglebutton containing a miniature of the given (standard) page type
- *      The toggle button is initialized to follow a GAction with the given action name and target value entryNb.
- *      The returned widget is a floating ref.
- */
-GtkWidget* createEntryWithPreview(const PageTypeInfo* pti, size_t entryNb, const std::string_view& prefixedActionName) {
-    GtkWidget* button = gtk_toggle_button_new();
-
-    // // Use to restore labels in the menu
-    // GtkWidget* label = gtk_box_new(GTK_ORIENTATION_VERTICAL, 4);
-    // gtk_box_append(GTK_BOX(label), createPreviewImage(*pti));
-    // gtk_box_append(GTK_BOX(label), gtk_label_new(pti->name.c_str()));
-    // gtk_button_set_child(GTK_BUTTON(button), label);
-
-    GtkWidget* preview = xoj::helper::createPreviewImage(pti->page);
-    gtk_widget_set_size_request(preview, xoj::helper::PREVIEW_WIDTH, xoj::helper::PREVIEW_HEIGHT);
-    gtk_widget_set_tooltip_text(preview, pti->name.c_str());
-    gtk_button_set_child(GTK_BUTTON(button), preview);  // takes ownership of preview
-    gtk_actionable_set_action_name(GTK_ACTIONABLE(button), prefixedActionName.data());
-    gtk_actionable_set_action_target_value(GTK_ACTIONABLE(button), xoj::util::makeGVariantSPtr(entryNb).get());
-    return button;
-}
-
-/**
- * @brief Create a grid containing a miniature for each one the given (standard) page types
- *      The returned widget is a floating ref.
- */
-GtkWidget* createPreviewGrid(const std::vector<std::unique_ptr<PageTypeInfo>>& pageTypes,
-                             const std::string_view& prefixedActionName) {
-    GtkGrid* grid = createEmptyGrid();
-    gtk_widget_set_margin_top(GTK_WIDGET(grid), 10);
-
-    size_t n = 0;
-    int gridX = 0;
-    int gridY = 0;
-
-    for (const auto& pageInfo: pageTypes) {
-        // Special page types do not get a preview
-        xoj_assert(!pageInfo->page.isSpecial());
-        auto* entry = createEntryWithPreview(pageInfo.get(), n++, prefixedActionName);
-        if (gridX >= PAGE_TYPES_PER_ROW) {
-            gridX = 0;
-            gridY++;
-        }
-        gtk_grid_attach(grid, entry, gridX, gridY, 1, 1);
-        gridX++;
-    }
-    return GTK_WIDGET(grid);
-}
-
 GtkWidget* createOrientationButton(std::string_view actionName, GtkOrientation orientation, std::string_view icon) {
     GtkWidget* button = gtk_toggle_button_new();
     gtk_button_set_icon_name(GTK_BUTTON(button), icon.data());
@@ -143,7 +82,6 @@ GtkWidget* createSeparator() {
     gtk_widget_set_margin_end(separator, 7);
     return separator;
 }
-
 };  // namespace
 
 PageTypeSelectionPopover::PageTypeSelectionPopover(Control* control, Settings* settings, GtkApplicationWindow* win):
@@ -242,13 +180,14 @@ GtkWidget* PageTypeSelectionPopover::createPopover() const {
     GtkBox* box = GTK_BOX(gtk_box_new(GTK_ORIENTATION_VERTICAL, 0));
     gtk_popover_set_child(GTK_POPOVER(popover), GTK_WIDGET(box));
 
-    GtkWidget* previewGrid = createPreviewGrid(types->getPageTypes(), prefixedPageTypeActionName);
+    GtkWidget* previewGrid =
+            xoj::helper::PageTypeGrid::createPreviewGrid(types->getPageTypes(), prefixedPageTypeActionName);
     gtk_widget_set_margin_bottom(previewGrid, 5);
     gtk_box_append(box, previewGrid);
 
     gtk_box_append(box, createSeparator());
 
-    GtkGrid* specialTypesGrid = createEmptyGrid();
+    GtkGrid* specialTypesGrid = xoj::helper::PageTypeGrid::createEmptyGrid();
     gtk_widget_set_margin_top(GTK_WIDGET(specialTypesGrid), 5);
     gtk_widget_set_margin_bottom(GTK_WIDGET(specialTypesGrid), 5);
 
@@ -267,7 +206,7 @@ GtkWidget* PageTypeSelectionPopover::createPopover() const {
                         1, 1);
     }
 
-    while (gridX < PAGE_TYPES_PER_ROW) {
+    while (gridX < xoj::helper::PageTypeGrid::PAGE_TYPES_PER_ROW) {
         // Add empty cells to the grid so the buttons don't get extended
         gtk_grid_attach(specialTypesGrid, gtk_label_new(""), gridX++, 0, 1, 1);
     }
